ctz-mailer/app/requirements.txt
Django==3.2.6
djangorestframework
django-cors-headers
djangorestframework-jwt
psycopg2-binary
gunicorn
flake8
six
celery
